refactor: Extract and reorganize test modules for better clarity and maintainability
