+++ src/range/dynamic.rs
@@ -308,3 +308,191 @@ impl<'de> serde::de::Deserialize<'de> for DynamicVersionRange {
         })
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::Comparator;
+    use crate::VersError;
+    use crate::range::VersionRange;
+    use super::DynamicVersionRange;
+
+    #[test]
+    fn test_parse_simple() {
+        let range: DynamicVersionRange = "vers:npm/1.2.3".parse().unwrap();
+        assert_eq!(range.versioning_scheme(), "npm");
+        assert_eq!(range.constraints().len(), 1);
+        assert_eq!(range.constraints()[0].comparator, Comparator::Equal);
+        assert_eq!(range.constraints()[0].version.to_string(), "1.2.3");
+    }
+
+    #[test]
+    fn test_parse_with_comparators() {
+        let range: DynamicVersionRange = "vers:npm/>=1.0.0|<2.0.0".parse().unwrap();
+        assert_eq!(range.versioning_scheme(), "npm");
+        assert_eq!(range.constraints().len(), 2);
+        assert_eq!(
+            range.constraints()[0].comparator,
+            Comparator::GreaterThanOrEqual
+        );
+        assert_eq!(range.constraints()[0].version.to_string(), "1.0.0");
+        assert_eq!(range.constraints()[1].comparator, Comparator::LessThan);
+        assert_eq!(range.constraints()[1].version.to_string(), "2.0.0");
+    }
+
+    #[test]
+    fn test_parse_star() {
+        let range: DynamicVersionRange = "vers:npm/*".parse().unwrap();
+        assert_eq!(range.versioning_scheme(), "npm");
+        assert_eq!(range.constraints().len(), 1);
+        assert_eq!(range.constraints()[0].comparator, Comparator::Any);
+        assert_eq!(range.constraints()[0].version.to_string(), "0.0.0");
+    }
+
+    #[test]
+    fn test_parse_with_spaces() {
+        let range: DynamicVersionRange = "vers:npm/ >= 1.0.0 | < 2.0.0 ".parse().unwrap();
+        assert_eq!(range.versioning_scheme(), "npm");
+        assert_eq!(range.constraints().len(), 2);
+        assert_eq!(
+            range.constraints()[0].comparator,
+            Comparator::GreaterThanOrEqual
+        );
+        assert_eq!(range.constraints()[0].version.to_string(), "1.0.0");
+        assert_eq!(range.constraints()[1].comparator, Comparator::LessThan);
+        assert_eq!(range.constraints()[1].version.to_string(), "2.0.0");
+    }
+
+    #[test]
+    fn test_parse_with_url_encoding() {
+        let range: DynamicVersionRange = "vers:npm/1.0.0%2Bbuild.1".parse().unwrap();
+        assert_eq!(range.versioning_scheme(), "npm");
+        assert_eq!(range.constraints().len(), 1);
+        assert_eq!(range.constraints()[0].comparator, Comparator::Equal);
+        assert_eq!(range.constraints()[0].version.to_string(), "1.0.0+build.1");
+    }
+
+    #[test]
+    fn test_invalid_constraint_simplification() {
+        let result: DynamicVersionRange = "vers:npm/1.2.3|<2.0.0".parse().unwrap();
+        assert_eq!(result.to_string(), "vers:npm/<2.0.0");
+
+        let result: DynamicVersionRange = "vers:npm/>1.0.0|>2.0.0".parse().unwrap();
+        assert_eq!(result.to_string(), "vers:npm/>1.0.0");
+
+        let result: DynamicVersionRange = "vers:npm/<1.0.0|<2.0.0".parse().unwrap();
+        assert_eq!(result.to_string(), "vers:npm/<2.0.0");
+    }
+
+    #[test]
+    fn test_display() {
+        let range: DynamicVersionRange = "vers:npm/>=1.0.0|<2.0.0".parse().unwrap();
+        assert_eq!(range.to_string(), "vers:npm/>=1.0.0|<2.0.0");
+
+        let range: DynamicVersionRange = "vers:npm/*".parse().unwrap();
+        assert_eq!(range.to_string(), "vers:npm/*");
+
+        let range: DynamicVersionRange = "vers:npm/1.2.3".parse().unwrap();
+        assert_eq!(range.to_string(), "vers:npm/1.2.3");
+    }
+
+    #[test]
+    fn test_dynamic_parse_npm() {
+        let range: DynamicVersionRange = "vers:npm/>=1.0.0|<2.0.0".parse().unwrap();
+        assert_eq!(range.versioning_scheme(), "npm");
+        assert_eq!(range.constraints().len(), 2);
+    }
+
+    #[test]
+    fn test_dynamic_parse_semver() {
+        let range: DynamicVersionRange = "vers:semver/>=1.0.0|<2.0.0".parse().unwrap();
+        assert_eq!(range.versioning_scheme(), "semver");
+        assert_eq!(range.constraints().len(), 2);
+    }
+
+    #[test]
+    fn test_dynamic_parse_unsupported() {
+        let range: Result<DynamicVersionRange, VersError> = "vers:pypi/>=1.0.0|<2.0.0".parse();
+        assert!(range.is_err());
+        assert!(matches!(
+            range.unwrap_err(),
+            VersError::UnsupportedVersioningScheme(_)
+        ));
+    }
+
+    #[test]
+    fn test_dynamic_contains() {
+        let range: DynamicVersionRange = "vers:npm/>=1.0.0|<2.0.0".parse().unwrap();
+        assert!(range.contains("1.5.0".to_string()).unwrap());
+        assert!(!range.contains("2.0.0".to_string()).unwrap());
+        assert!(!range.contains("0.9.0".to_string()).unwrap());
+    }
+
+    #[test]
+    fn test_dynamic_contains_invalid_version() {
+        let range: DynamicVersionRange = "vers:npm/>=1.0.0|<2.0.0".parse().unwrap();
+        let result = range.contains("invalid.version".to_string());
+        assert!(result.is_err());
+        assert!(matches!(
+            result.unwrap_err(),
+            VersError::InvalidVersionFormat(..)
+        ));
+    }
+
+    #[test]
+    fn test_dynamic_display() {
+        let range: DynamicVersionRange = "vers:npm/>=1.0.0|<2.0.0".parse().unwrap();
+        assert_eq!(range.to_string(), "vers:npm/>=1.0.0|<2.0.0");
+    }
+
+    #[test]
+    fn test_dynamic_equality() {
+        let range1: DynamicVersionRange = "vers:npm/>=1.0.0|<2.0.0".parse().unwrap();
+        let range2: DynamicVersionRange = "vers:npm/>=1.0.0|<2.0.0".parse().unwrap();
+        let range3: DynamicVersionRange = "vers:semver/>=1.0.0|<2.0.0".parse().unwrap();
+
+        assert_eq!(range1, range2);
+        // Both should parse to the same SemVer range
+        assert_eq!(range1.constraints(), range3.constraints());
+    }
+
+    #[test]
+    fn test_contains_simple() {
+        let range: DynamicVersionRange = "vers:npm/1.2.3".parse().unwrap();
+        assert!(range.contains("1.2.3".to_string()).unwrap());
+        assert!(!range.contains("1.2.4".to_string()).unwrap());
+    }
+
+    #[test]
+    fn test_contains_range() {
+        let range: DynamicVersionRange = "vers:npm/>=1.0.0|<2.0.0".parse().unwrap();
+        assert!(range.contains("1.0.0".to_string()).unwrap());
+        assert!(range.contains("1.5.0".to_string()).unwrap());
+        assert!(!range.contains("2.0.0".to_string()).unwrap());
+        assert!(!range.contains("0.9.0".to_string()).unwrap());
+    }
+
+    #[test]
+    fn test_contains_star() {
+        let range: DynamicVersionRange = "vers:npm/*".parse().unwrap();
+        assert!(range.contains("1.0.0".to_string()).unwrap());
+        assert!(range.contains("2.0.0".to_string()).unwrap());
+        assert!(range.contains("0.0.1".to_string()).unwrap());
+    }
+
+    #[test]
+    fn test_contains_not_equal() {
+        let range: DynamicVersionRange = "vers:npm/!=1.2.3".parse().unwrap();
+        assert!(!range.contains("1.2.3".to_string()).unwrap());
+        assert!(range.contains("1.2.4".to_string()).unwrap());
+    }
+
+    #[test]
+    fn test_contains_complex() {
+        let range: DynamicVersionRange = "vers:npm/>=1.0.0|<2.0.0|!=1.5.0".parse().unwrap();
+        assert!(range.contains("1.0.0".to_string()).unwrap());
+        assert!(range.contains("1.7.0".to_string()).unwrap());
+        assert!(!range.contains("1.5.0".to_string()).unwrap());
+        assert!(!range.contains("2.0.0".to_string()).unwrap());
+        assert!(!range.contains("0.9.0".to_string()).unwrap());
+    }
+}
